+++ src/Resources/ResourcesManger.cpp
@@ -274,7 +274,7 @@ bool ResourceManger::loadJSONResources(const std::string& JSONPath)
 							for (const auto& currentFrame : framesArray)
 							{
 								const std::string subTextureStr = currentFrame["subTexture"].GetString();
-								const uint64_t duration = currentFrame["duration"].GetInt64();
+								const double duration = currentFrame["duration"].GetDouble();
 								const auto pTextureAtlas = getTexture(textureAtlas);
 								const auto pSubTexture = pTextureAtlas->getSubTexture(subTextureStr);
 								framesDescroptions.emplace_back(pSubTexture.leftBottomUV, pSubTexture.rightTopUV, duration);
